refactor(port_and_speed): extract speed test components - Extract  and  to separate files - Reduce  complexity (840 -> 246 lines) - Improve separation of concerns in dashboard components
